+++ app/upgrades.go
@@ -5,6 +5,8 @@ import (
 
 	icqtypes "github.com/strangelove-ventures/async-icq/v6/types"
 
+	"cosmossdk.io/math"
+
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -110,6 +112,7 @@ var upgrades = map[string]appUpgrade{
 
 			removeInactiveValidatorDelegations(ctx, app)
 			setupICQ(ctx, app)
+			updateMaxSupply(ctx, app)
 
 			return vm, nil
 		},
@@ -128,6 +131,7 @@ var upgrades = map[string]appUpgrade{
 
 			removeInactiveValidatorDelegations(ctx, app)
 			setupICQ(ctx, app)
+			updateMaxSupply(ctx, app)
 
 			return vm, nil
 		},
@@ -325,4 +329,14 @@ func setupICQ(ctx sdk.Context, app *App) {
 	ctx.Logger().Info("Updating ICQ params")
 	app.ICQKeeper.SetParams(ctx, icqtypes.NewParams(true, []string{"/provenance.oracle.v1.Query/Oracle"}))
 	ctx.Logger().Info("Done updating ICQ params")
+}
+
+// updateMaxSupply sets the value of max supply to the current value of MaxTotalSupply
+func updateMaxSupply(ctx sdk.Context, app *App) {
+	ctx.Logger().Info("Updating MaxSupply marker param")
+	params := app.MarkerKeeper.GetParams(ctx)
+	//nolint:staticcheck // Populate new param with deprecated param
+	params.MaxSupply = math.NewIntFromUint64(params.MaxTotalSupply)
+	app.MarkerKeeper.SetParams(ctx, params)
+	ctx.Logger().Info("Done updating MaxSupply marker param")
 }